+++ RevenueCatUI/UIKit/PaywallViewController.swift
@@ -27,13 +27,13 @@ public class PaywallViewController: UIViewController {
     /// See ``PaywallViewControllerDelegate`` for receiving purchase events.
     @objc public final weak var delegate: PaywallViewControllerDelegate?
 
-    var mode: PaywallViewMode {
-        return .fullScreen
+    private var configuration: PaywallViewConfiguration {
+        didSet {
+            // Overriding the configuration requires re-creating the `HostingViewController`.
+            // This is used by some Hybrid SDKs that require modifying the content after creation.
+            self.hostingController = self.createHostingController()
+        }
     }
-
-    private let offering: Offering?
-    private let fonts: PaywallFontProvider
-    private let displayCloseButton: Bool
 
     /// Initialize a `PaywallViewController` with an optional `Offering`.
     /// - Parameter offering: The `Offering` containing the desired `PaywallData` to display.
@@ -54,16 +54,47 @@ public class PaywallViewController: UIViewController {
     /// Initialize a `PaywallViewController` with an optional `Offering` and ``PaywallFontProvider``.
     /// - Parameter offering: The `Offering` containing the desired `PaywallData` to display.
     /// `Offerings.current` will be used by default.
-    /// - Parameter fonts: An optional ``PaywallFontProvider``.
+    /// - Parameter fonts: A ``PaywallFontProvider``.
     /// - Parameter displayCloseButton: Set this to `true` to automatically include a close button.
-    public init(
+    public convenience init(
         offering: Offering? = nil,
         fonts: PaywallFontProvider,
         displayCloseButton: Bool = false
     ) {
-        self.offering = offering
-        self.fonts = fonts
-        self.displayCloseButton = displayCloseButton
+        self.init(
+            content: .optionalOffering(offering),
+            fonts: fonts,
+            displayCloseButton: displayCloseButton
+        )
+    }
+
+    /// Initialize a `PaywallViewController` with an offering identifier.
+    /// - Parameter offeringIdentifier: The identifier for the offering with `PaywallData` to display.
+    /// - Parameter fonts: A ``PaywallFontProvider``.
+    /// - Parameter displayCloseButton: Set this to `true` to automatically include a close button.
+    public convenience init(
+        offeringIdentifier: String,
+        fonts: PaywallFontProvider = DefaultPaywallFontProvider(),
+        displayCloseButton: Bool = false
+    ) {
+        self.init(
+            content: .offeringIdentifier(offeringIdentifier),
+            fonts: fonts,
+            displayCloseButton: displayCloseButton
+        )
+    }
+
+    init(
+        content: PaywallViewConfiguration.Content,
+        fonts: PaywallFontProvider,
+        displayCloseButton: Bool
+    ) {
+        self.configuration = .init(
+            content: content,
+            mode: Self.mode,
+            fonts: fonts,
+            displayCloseButton: displayCloseButton
+        )
 
         super.init(nibName: nil, bundle: nil)
     }
@@ -73,55 +104,10 @@ public class PaywallViewController: UIViewController {
         fatalError("init(coder:) has not been implemented")
     }
 
-    private lazy var hostingController: UIHostingController<some View> = {
-        let view = PaywallView(offering: self.offering,
-                               customerInfo: nil,
-                               mode: self.mode,
-                               fonts: self.fonts,
-                               displayCloseButton: self.displayCloseButton,
-                               introEligibility: nil,
-                               purchaseHandler: nil)
-            .onPurchaseCompleted { [weak self] transaction, customerInfo in
-                guard let self else { return }
-                self.delegate?.paywallViewController?(self, didFinishPurchasingWith: customerInfo)
-                self.delegate?.paywallViewController?(self,
-                                                      didFinishPurchasingWith: customerInfo,
-                                                      transaction: transaction)
-            }
-            .onPurchaseCancelled { [weak self] in
-                guard let self else { return }
-                self.delegate?.paywallViewControllerDidCancelPurchase?(self)
-            }
-            .onRestoreCompleted { [weak self] customerInfo in
-                guard let self else { return }
-                self.delegate?.paywallViewController?(self, didFinishRestoringWith: customerInfo)
-            }
-            .onSizeChange { [weak self] in
-                guard let self else { return }
-                self.delegate?.paywallViewController?(self, didChangeSizeTo: $0)
-            }
-
-        return .init(rootView: view)
-    }()
-
     public override func loadView() {
         super.loadView()
 
-        self.addChild(self.hostingController)
-        self.view.addSubview(self.hostingController.view)
-        self.hostingController.didMove(toParent: self)
-        self.hostingController.view.translatesAutoresizingMaskIntoConstraints = false
-
-        NSLayoutConstraint.activate([
-            self.hostingController.view.topAnchor.constraint(equalTo: view.topAnchor),
-            self.hostingController.view.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-            self.hostingController.view.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            self.hostingController.view.trailingAnchor.constraint(equalTo: view.trailingAnchor)
-        ])
-
-        // make the background of the container clear so that if there are cutouts, they don't get
-        // overridden by the hostingController's view's background.
-        self.hostingController.view.backgroundColor = .clear
+        self.hostingController = self.createHostingController()
     }
 
     public override func viewDidDisappear(_ animated: Bool) {
@@ -130,7 +116,55 @@ public class PaywallViewController: UIViewController {
         }
         super.viewDidDisappear(animated)
     }
+
+    /// - Warning: For internal use only
+    @objc(updateWithOffering:)
+    public func update(with offering: Offering) {
+        self.configuration.content = .offering(offering)
+    }
+
+    /// - Warning: For internal use only
+    @objc(updateWithOfferingIdentifier:)
+    public func update(with offeringIdentifier: String) {
+        self.configuration.content = .offeringIdentifier(offeringIdentifier)
+    }
+
+    // MARK: - Internal
+
+    class var mode: PaywallViewMode {
+        return .fullScreen
+    }
+
+    // MARK: - Private
+
+    private var hostingController: UIHostingController<PaywallContainerView>? {
+        willSet {
+            guard let oldValue = self.hostingController else { return }
+
+            oldValue.willMove(toParent: nil)
+            oldValue.view.removeFromSuperview()
+            oldValue.removeFromParent()
+        }
+
+        didSet {
+            guard let newController = self.hostingController else { return }
+
+            self.addChild(newController)
+            self.view.addSubview(newController.view)
+            newController.didMove(toParent: self)
+
+            NSLayoutConstraint.activate([
+                newController.view.topAnchor.constraint(equalTo: self.view.topAnchor),
+                newController.view.bottomAnchor.constraint(equalTo: self.view.bottomAnchor),
+                newController.view.leadingAnchor.constraint(equalTo: self.view.leadingAnchor),
+                newController.view.trailingAnchor.constraint(equalTo: self.view.trailingAnchor)
+            ])
+        }
+    }
+
 }
+
+// MARK: - PaywallViewControllerDelegate
 
 /// Delegate for ``PaywallViewController``.
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, *)
@@ -169,6 +203,67 @@ public protocol PaywallViewControllerDelegate: AnyObject {
     @objc(paywallViewController:didChangeSizeTo:)
     optional func paywallViewController(_ controller: PaywallViewController,
                                         didChangeSizeTo size: CGSize)
+
+}
+
+// MARK: - Private
+
+@available(iOS 15.0, macOS 12.0, tvOS 15.0, *)
+private extension PaywallViewController {
+
+    func createHostingController() -> UIHostingController<PaywallContainerView> {
+        let container = PaywallContainerView(
+            configuration: self.configuration,
+            purchaseCompleted: { [weak self] transaction, customerInfo in
+                guard let self else { return }
+                self.delegate?.paywallViewController?(self, didFinishPurchasingWith: customerInfo)
+                self.delegate?.paywallViewController?(self,
+                                                      didFinishPurchasingWith: customerInfo,
+                                                      transaction: transaction)
+            },
+            purchaseCancelled: { [weak self] in
+                guard let self else { return }
+                self.delegate?.paywallViewControllerDidCancelPurchase?(self)},
+            restoreCompleted: { [weak self] customerInfo in
+                guard let self else { return }
+                self.delegate?.paywallViewController?(self, didFinishRestoringWith: customerInfo)
+            },
+            onSizeChange: { [weak self] in
+                guard let self else { return }
+                self.delegate?.paywallViewController?(self, didChangeSizeTo: $0)
+            }
+        )
+
+        let controller = UIHostingController(rootView: container)
+
+        // make the background of the container clear so that if there are cutouts, they don't get
+        // overridden by the hostingController's view's background.
+        controller.view.backgroundColor = .clear
+        controller.view.translatesAutoresizingMaskIntoConstraints = false
+
+        return controller
+    }
+
+}
+
+@available(iOS 15.0, macOS 12.0, tvOS 15.0, *)
+private struct PaywallContainerView: View {
+
+    var configuration: PaywallViewConfiguration
+
+    let purchaseCompleted: PurchaseCompletedHandler
+    let purchaseCancelled: PurchaseCancelledHandler
+    let restoreCompleted: PurchaseOrRestoreCompletedHandler
+    let onSizeChange: (CGSize) -> Void
+
+    var body: some View {
+        PaywallView(configuration: self.configuration)
+            .onPurchaseCompleted(self.purchaseCompleted)
+            .onPurchaseCancelled(self.purchaseCancelled)
+            .onRestoreCompleted(self.restoreCompleted)
+            .onSizeChange(self.onSizeChange)
+
+    }
 
 }