+++ Tests/APITesters/RevenueCatUIAPITester/SwiftAPITester/PaywallViewControllerAPI.swift
@@ -18,6 +18,7 @@ func paywallViewControllerAPI(_ delegate: Delegate, _ offering: Offering?) {
 
     let _: UIViewController = PaywallViewController(fonts: fontProvider)
     let _: UIViewController = PaywallViewController(offering: offering)
+    let _: UIViewController = PaywallViewController(offeringIdentifier: "offering")
     let _: UIViewController = PaywallViewController(displayCloseButton: true)
     let _: UIViewController = PaywallViewController(fonts: fontProvider)
     let _: UIViewController = PaywallViewController(offering: offering, displayCloseButton: true)
@@ -25,6 +26,14 @@ func paywallViewControllerAPI(_ delegate: Delegate, _ offering: Offering?) {
     let _: UIViewController = PaywallViewController(offering: offering,
                                                     fonts: fontProvider,
                                                     displayCloseButton: true)
+    let _: UIViewController = PaywallViewController(offeringIdentifier: "offering", displayCloseButton: true)
+    let _: UIViewController = PaywallViewController(offeringIdentifier: "offering", fonts: fontProvider)
+    let _: UIViewController = PaywallViewController(offeringIdentifier: "offering",
+                                                    fonts: fontProvider,
+                                                    displayCloseButton: true)
+
+    controller.update(with: offering!)
+    controller.update(with: "offering_identifier")
 }
 
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, *)
@@ -33,6 +42,7 @@ func paywallFooterViewControllerAPI(_ delegate: Delegate, _ offering: Offering?)
     controller.delegate = delegate
 
     let _: UIViewController = PaywallFooterViewController(offering: offering)
+    let _: UIViewController = PaywallFooterViewController(offeringIdentifier: "offering")
 }
 
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, *)