+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 4.32.2
+### Other Changes
+* `PaywallViewController`: methods for reconfiguring paywall with new offering (#3592) via NachoSoto (@NachoSoto)
+* `Integration Tests`: verify `PaywallData` images can be loaded (#3596) via NachoSoto (@NachoSoto)
+* Simplify `CocoapodsInstallation` `Podfile` (#3593) via NachoSoto (@NachoSoto)
+## 4.32.1
+### RevenueCatUI
+* `PaywallViewController`: new initializer with `Offering` identifier (#3587) via NachoSoto (@NachoSoto)
+* `Paywalls`: improve template 5 layout for long product names (#3589) via NachoSoto (@NachoSoto)
+### Other Changes
+* `Paywalls`: extracted `PaywallViewConfiguration` (#3586) via NachoSoto (@NachoSoto)
+* `CircleCI`: avoid installing `Xcodes` when not needed (#3585) via NachoSoto (@NachoSoto)
+* `CircleCI`: change all jobs to M1 (#3140) via NachoSoto (@NachoSoto)
+
 ## 4.32.0
 ### New Features
 * `StoreProduct`: add localized price per period strings (#3546) via Andy Boedo (@aboedo)